git-svn-id: https://minsktranssched.svn.sourceforge.net/svnroot/minsktranssched/trunk@7 1edd68fb-5641-41d5-9a6f-76477202bf19

+++ ScheduleConverter/src/BusStop.java
@@ -1,7 +1,7 @@
 
 public class BusStop
 {
-	short id;
+	int id;
 	String name;
 	String officialName;
 	String description;
